Replace expensive pictures-by-type with stats computation
Currently, computing folder/repository stats is done by expensive
operations: multiple, repeated iterations over the folder pictures,
each time testing for a given picture type. This is of course
expensive, and it's much cheaper to iterate only once, summing stats
as we go along.

The folder class stats is updated as well, to take place during
picture stats computation, so in the end for getting both the picture
and folder class stats we test each picture in the repository only
once for its type.

This results in quite good performance, and the next step will be to
improve per-picture checks with something cheaper.