Add support for standalone panos (ranged) files
Currently, even if the raw files that were used to build a range
(e.g. panorama) file exists, if the range file is only present as a
processed file, it will be marked as standalone. By remembering range
information for a file and doing another reprocessing round at the
end, we can actually mark such ranged files as also having a raw
component, if the first member of the range exists.