Change unprocessed folder classification
Currently, a folder is 'unprocessed' only if it has some processed and some raw
files. In practice, this is suboptimal since:

- it makes harder to get to a list of all folders with unprocessed files;
  adding 'mixed' folders pulls in folders that are mixed not necessarily due to
  unprocessed files
- the 'unprocessed' folder view (raw,unprocessed) shows diferent totals than
  the homepage count of unprocessed files

The patch changes the classification such that a unprocessed folder is any
folder that has some processed and some unprocessed files, irrespective of
whether it also has (or not) standalone or orphaned files.