Change outdated file classification
Currently, the outdated classification is very pessimistic: it only uses mtime
for checks. However, for the (usual) case where exported processed files are
set to the original capture mtime, this mens that the sidecar file will
necessarily be always older, hence they'll always be considered orphaned.

This patch changes the classification to be more optimistic, likely generating
more false negatives: for processed files, take the max of mtime and ctime.
This will reduce the noise of outdated files, but it will also misclassify
unrelated inode changes as 'refreshing' the file time.

Given that the a correct algorithm is only possible by parsing exif metadata,
and that outdated is a 'weak' signal, the new model is cleaner/better.