Better support for images in sub folders
Currently, all images in a folder, whether split in sub folders or
directly under the folder, are squashed together in a single
level. This has two downsides:

- the sub folder information is lost, and thus the physical file paths
  shown in the image view is wrong, not matching the filesystem
- not able to handle the same name in different sub directories

This patch fixes both issues by using the full "sub-path" as the name
of the image. The only downside is that the folder image list is more
noisy.