Revert/improve on "Change outdated file classification"
This patch reverts the implementation of the above change and replaces
it with explicit storing of file ctime and only computing the maximum
at runtime, in mkImageStatus. This fixes the timeline, where the time
of jpegs is now mtime, and we only use ctime for outdated-vs-processed
classification.

Both methods are wrong, and the proper fix will be exif parsing.