Update scaffolding to Yesod 1.4
This is a huge change with lots of churn due to update from some very
old Yesod version to 1.4. The most significant user-visible change is
the redoing of the config file with only one section, containing
variables that can be overridden by environment, instead of separate
devel/test/production sections.