Switch to storing external resources in the repository
This commit changes the way external resources are handled:
previously, they were required to be provided by the user under a
given url. While this avoids any licence issues, it puts the work on
the user to configure a separate web server with specific
paths. Additionally, an external web server can't take advantage of
the way Yesod deals with static files: serving them with never-expire
headers, etc.

The new model is to embed the copies of files needed in this
repository; this is a more complex licensing model, which is detailed
in the LICENCE.md file, and might also have technical impact (security
updates of JavaScript files vs. testing of all components together).

New external files: jquery, jquery-metadata, jquery-tablesorter, bootstrap.