Remove dependency and use of classyprelude-yesod
While looking for something else, a strace showed that the binary was
opening all files from `/etc/ssl/certs`; this was unexpected, since
the program doesn't do any client HTTP requests. Upon investigation,
this is due to the use of ClassyPrelude.Yesod, which depends on
Network.HTTP.Client.

Since we don't need this, drop the classyprelude-yesod, which brings
the non-stripped binary from 33MiB to 26MiB (!).