Add navigation buttons between images in a folder
I don't like the placement of the buttons, but after playing with it
for a while this is the best I could find (in the Basic info panel
it's out of place, as a separate line it's noise). The code is fine
anyway, might just need to relocate the buttons later.