Introduce image flags and track soft master status
Add an image flags data type which (for now) tracks the soft master
status. This allows the hard master file (if any) to be chosen over a
soft master one (which is demoted to a jpeg file), which restores
"mixed tree" behaviour for source dirs and allows proper image
tracking again.