Switch build system to stack and hpack
Newer yesod released have switched to stack and hpack, which is a very
good thing, as the stackage LTS releases are a much better way to
manage deps. The only downside is the need to have network access
during build/development (at least initially).

The hpack move is more for fun, rather than a serious need. Will see
in time if it helps.