Redo how main div positioning is done
Instead of doing it via current top position, compute position based
on navbar height. This should, in principle, be the same; however, it
has the advantage that it allows easy reset of the position in case
the layout changes (e.g. narrower or wider screen, which leads to
navbar resize, which means we have to re-position change), as opposed
to resetting the div to position:relative, getting the offset back,
and then switching again to fixed.

This actually allows consistent behaviour in case of orientation
changes on mobile.