Improve image loading messages
Currently, a message with picture name is written as soon as we
schedule the image draw. However, this can be delayed (per the
somewhat recent 'delay draw if not loaded' change), which causes the
UI to get confusing with regards to what is the state - loading, or…?

Fix this by adding a "loading" message, and only draw image name after
it has been put on the scren.