Implement downscaling of images for optimised viewing
Not happy with this approach, but downloading and scaling a 13MB image
on a tablet is a suboptimal user experience.

The approach is as follows:

- add configuration items for which sizes we allow; when a client
  request a specific size, we find the smallest size in our lis that
  is greater or equal than the one requested by the client
- generate and cache the image
- serve it back to the client

Additionally, a list of sizes is pre-generated (in the background) as
soon as the application first scans the filesystem, which might take
days on a larget collection (and depending on which sizes are
requested).