Switch thumbnails to transparent, square PNGs
This improves (IMHO) significantly the folder view, since the
alignment is better in the thumbnail column.

The new code still uses hardcoded image/jpeg in some cases, so that's
that, but for generated images (thumbnails/scaled images), it matche
the explicit format we pass to ImageMagick (thus one step better).