More strictness…
This is almost to the level of "use Strict", but I try to keep to the
laziness ideal.

The problem is that without merging, even with strict Maps, the values
don't seem to be forced, and thus folder merging lags
significantly (sometimes). It is easily viewable by retainer
profiling or by scanner timing stats.