Exif: introduce location/people/keyword parsing
I don't know if HierarchicalSubject is a standard thing or
Lightroom-specific, but it works for my use case.

Of course, during this exercise it turned out that Location being a Maye
is wrong, since hierarchical locations mean more than one location even
for a single picture (or file, for that matter).