Fixup parsing of create date multiple fields
The msum here runs in the parser monad, not in the Maybe monad. So
therefore a parse of Nothing is successful, as the requested fields
are optional. Switch to required so that the fields are checked in
order as expected, and apply a graceful fall back around the entire
parse call, so that we don't fail the entire parsing if there are no
date fields.

At least, that's what I understand about how this all interacts :/