Switch from PCRE to TDFA as regex library
I just learned that PCRE is tricky with regards to non-ASCII text, and
it's not entirely clear what all the needed options are to make it
work. So let's simply switch to TDFA, since it's native and works with
it by default.

This still doesn't fix all the underlying copy regex issues though.