Refresh previews and thumbnails also on ctime changes
When a raw-to-processed version export sets the final file mtime to
capture date, refreshing previews on mtime only will not trigger the
needed rebuild. On the downside, tying in with ctime will lead to
extra rebuilds since e.g. permission changes will also trigger a
rebuild. But it is the more correct behaviour, so let's do that.