Implement partial error handling in image view
While trying something else, I accidentally learned about blaze-svg,
and realized that it would be possible to construct on-the-fly
responses to images when there is a backend error. Basically, for the
viewer, we shouldn't ever send a non-2xx response, and that means we
always have to provide a valid image. So let's build one :)