Record both raw and parsed exif values
This is an incompatible change in the exif output: instead of just
raw, use both raw and parsed values, by using 'exiftool -json
-l'. Since the raw cache now differs from stable version (and as this
could happen in the future again), switch dev raw cache to its own
file.

Additionally, switch the lens gathering to use the parsed values, as
this was the main impetus for the change.