Track location-related exif fields separately
Currently, they're all lumped into exifLocation, which erases semantic
differences (country vs. city vs. actual location). This makes, for
example, the home page cluttered - countries will always be much
higher in the ranking.

So move to track and promote them separately. The promotion is not
perfect, as we lose the upwards chain when e.g. merging cities, so
there is work left to do here.

This move also requires dropping the use of the hierarchical location
keyword tree, but that's fine, at least for my own use case.