Parallelise the directory scanning
On my picture trees, using a fully up-to-date cache (so no exif
rebuild), this improves the speed by more than 2.5× (for 8
cores). Very sublinear, but good nevertheless, plus improves the
parallel GC work balance (20%→~38%), at the cost of slight memory
increase (~15%).

On not up-to-date trees, this should improve speed-up much more, since
exiftool is running single-core, so the speed-ups could be getting
closer to linear in number of cores.

The parallelisation is two-level:

- we run the top-level dirs (all raws, all jpegs) in parallel
- we run the individual sub-dir scanning in parallel as well

This should result in high parallelism, still bounded by the number of
capabilities (I think).