Switch back to standard tablesorter bootstrap theme
I did not realize that simply defining a new theme means all the CSS
is lost - which resulted in no so nice e.g. filter-disabled
icons (very big and ugly).

Now the overrides are much more clear as well.

Also switch back to enabling the filter bar always, as 1) hovering on
mobile would be tricky, and 2) the hiding (via opacity) doesn't play
nice with local CSS.