Fixup panel links on homepage
Per bootstrap's documentation, “You may also include any <h1>-<h6>
with a .panel-title class to add a pre-styled heading. … For proper
link coloring, be sure to place links in headings within
.panel-title.”

OK, this explains the quite bad link appearance before, so let's fix
it. The "Year" heading needs the same treatment, even though it's not
a link, for consistency.