Change the meaning of folder searching
For a single (basic) atom, things remain the same. But for composite
atoms (e.g. x and y), the previous version meant "list all folder that
have either a picture with x and one with y", which is much less
helpful than "list all folders that have a picture with 'x and
y'". However, this is still broken end to end when implicitly
combining multiple atoms, since the check in getSearchFolderR has the
same error - checking the individual atoms separately.