Introduce negative atoms
This patch introduces the ability to say "nobody is in the picture",
or "has no country information", whereas the 'not' atom currently
available can only negate a positive assertion ('x is not in the
picture', 'country is not y').

The immediate approach of using empty values as negative cases (i.e.,
?person=, ?country) would work, but would lose the ability to search
for indeed defined but empty keyword/locations/etc., which sadly are
present sometimes in pictures. So instead, a special "no-" prefix is
used to denote such cases, which complicates a bit the parsing.

This new method can use improved atom descriptions, so do that as well
with better handling for empty cases as well.