gExit*: track "unknown" values as well
The split of keywords from Set Text to multiple Maybe Text introduced
a copy-paste bug: because Maybe is Foldable, foldl' works for it as
well, but all it does is ensure that we don't add Nothing values to
the map.

If we do want to track unknown values (as such, not as
"unknown"::Text), then addExifToGroup needs to be fixed, and as a
cascading effect, a number of UI elements need change as well.