First version of search operators
Besides the existing equality match, this adds fuzzy match and numeric
lt/gt. Of course, the main gain is the framework for operators, not
the existing ones.

However, not all is good. I couldn't make the type system play nice
enough to prevent (e.g.) a Person (OpLt x), only that x will not be
numeric. Not much gain, there. Will probably have to split the
operator type in two, and stop using GADTs.