Convert quick search to multi-word searches
Went back and forth on the exact behaviour that this should have, and
while not entirely happy with this version, it's at least consistent
across various searches: multiple meanings of the same word are
combined with Any, and all of these results are combined with
All. This way, no word is entirely ignored, although some searches
don't make sense - multiple words that don't match a Set-like
element (i.e. person/keyword right now).