Move the footer to default-layout
In the wrapper file, the footer follows pageBody, which includes the
scripts, so it is rendered after the scripts are loaded, leading to
flickering. The current Yesod scaffolding sites already has the footer
in the right place, not sure why I had it different.

Additionally, use .text-muted as a clean scaffolding site does, for
slightly better display - no need for the version to stand out.