Revert introduction of assets subdir and AssetsR
Sigh… the problem is that combineStatic, the function that Yesod uses
for combineStylesheets' (and thus the scaffolded site), doesn't
understand the concept of multiple static directories, so it doesn't
work on assets/ (as long as static/ is needed too). Thus, let's remove
the concept entirely and move files around (for fun).

On the other hand, when these files are served externally, having
multiple static directories is only more complication, so removing the
concept is not that bad.

But, churn is churn, sigh…