Switch to combineStylesheets for flicker-free page loads
With individual style sheets, the page "flickers" as the various CSS
load and the individual elements adjust size. The main offender is the
own basic.css that alters navbar properties (coming from
bootstrap.css), but even the simple fact of having the font awesome
style files that reference the fonts load separately from the rest of
the style sheets means a small one or two pixel vertical change.

With combined style sheets, a page reload (on Firefox) stays
completely stable - stable enough that previous commit that changed
the footer was needed, otherwise the footer was clearly rendering
async from the rest of the page.

The change in the font awesome .css files is safe, with or without
combining, as the combining post-process function will replace /static
with a proper relative reference (../), to account for the output css
file being in 'static/combined'.