Make a strong separation between release and dev builds
This change addresses two issues:

First, a number of things depend on the 'config/settings.yml'
file (hard-coded path!) as existing at build time; even worse, the
code won't build without this!  This is not good, as it makes the
build not quite reproducible (as I see it).

This issue is fixed by removing this build-dependency on a
configuration file, and instead:

- assume (at build time) that the static directory is 'static/', which
  makes sense as the VCS structure is fixed
- go even further, and remove more dependencies on "devel-like"
  settings such as mutable static, reload templates, etc. based on the
  build flag.

A special note is that combining is always used, since I find it
better behaved even for development builds.