Implement a fake full screen mode
Some browsers don't support a full screen mode. For these, implement a
fake full screen that simply hides the navbar and makes the background
black as well. It won't hide the browser's UI elements, but the navbar
can take a large part of the screen in some cases so at least that's
aside.