Introduce a special atom type ConstTrue
This atom is semantically equivalent to `all []`, but can be handled
specially in code to really mean any-and-all. It is true that `all []`
could be handled similarly, but it would mean specifically matching on
this value rather than constructor type (IMHO less nice).

One-liner fix for empty folder searches in next patch ☺