Switch filePath to a lazy Text
While it makes sense for the repository structure to be strict texts,
for the in-flight data (that goes away after a request), using lazy
texts should be better as no long-term storage for the chunks is
needed, and ideally some of the generated (lazy) texts can be very
cheap structures.

Of course, a lot of this is defeated by the repeated pack/unpack that
exist for now.