Rework how extension matching is performed
Technically, instead of keeping the extensions as a list of strings,
move to a Set Text, and match using the normal extension and not the
reversed one.

On an more abstract level, also move to only consider the last
extension, and not all. I think this makes more sense in any case.

While there's no measurable speed-up, this should be a bit faster,
based on micro-benchmarks.