Implement initial version of async scanning!
This is very bare-bones but seems to work (as tested with
further not-yet-committed logging code). It moves the scanning into a
separate thread (via the async library) and changes the rest of the
code to not require a valid repository as a result of the scanning
itself; this should make most "trigger scans" fast.

Most, because cancelling the previous thread seems to take somewhat
long in some cases; it could be due to the fact that we first launch
the new scan and only then cancel the previous one, leading to
concurrent (and wasted) work.