Fix flushing of search cache when the repository changes
Surprise, the minimal unit tests found a bug! Somewhere in the
refactoring of the scanning, the flushing of search cache was lost…
Fix it by flushing the cache before we attempt to update the
repository, even if we don't end up doing it, since the searches are
not that expensive; this means that at no time, the cache it
wrong (which could happen if we first updated and then flushed the
cache).

Also fix indentation in the updateRepo function, was very confusing.