Stop calling flushSearchCache from loadCachedOrScan
Since the flushing is done (automatically) from updateRepo, it doesn't
make sense to call it explicitly from the loadCachedOrScan, since all
modifications are anyway flushing the cache. Note: the comment was
obsolete already.

The removal of the flush cache and the use of view patterns allows
simplification of the code here, quite a bit, so let's do that as
well.