Change the icons for the status link
The "finished hourglass" that's now there for RepoFinished is very
confusing (to me), so while searching icons I came upon a new model: a
start hourglass for starting phase, a half-done hourglass for
scanning (which kind of makes sense), a check-mark for scanning
finished (rendering), a double check-mark for fully finished.

The only non-obvious thing is the transition from half-done hourglass
to check-mark, as there's no state that "finished hourglass" could be
used for. But overall this scheme is much more clear.