Fix bug in any/all atom handling
The any/all currently pop the entire stack, which means the some
things cannot actually be represented, e.g. (and (any a b c) (any c
d e)), because the second any parsing will turn this into (and (any a b
c (any c d e)), which is an error.

Fix this by encoding the amount of stack to pop in the value of the
any/all atom, which (fortunately?) had no value until now. This
encoding is applied on show and read on parsing.

This means a quick search of "notfound1 notfound2" leads correctly to
not found, instead of parse error.