Pre-compute and cache folder stats
This leads to a 10× improvement in things such as a folder list which
needs to compute stats for many folders. Additionally, the per-folder
stats were already computed in computeRepoStats, so the extra work is
at around 2× in throwing away intermediate folder stats (before
merging e.g. source and output folders).