Disable parallel GC in tests
This brings another 2× speedup, or even higher — ~3.3s to ~1.5s — so
very much worth doing. The test output is now very scrambled in the
SQL and logger output, but that's fine for usual runs, and it can be
fixed by `--ta -j1`, at least for the most part. Now test compilation
is much slower than runs :)