Rework parseCreateDate now that time library requires a timezone
Older time versions didn't actually require a timezone for '%Z', but
newer ones seem to (potentially since 1.9, "parsing: reject
invalid (and empty) time-zones with %z and %Z"). As such, the code
needs to try parsing simpler and simpler format strings.

This is just a quick fix. Now that time is actually well-behaved, the
code could be switched to ZonedTime instead, per the very old TODO.