Switch image name routes to multi-path
This changes URLs from '%2F' to actual (and proper) '/', which should
hopefully fix some crawl bot issues I've seen on demo.corydalis.io;
not sure if the errors are their side or mine, but from what I've
read, %2F encoding is not very nice (and needs special handling in the
web server, if reverse-proxying). The change itself is more than
trivial.