Fix viewer image resolution URL computation
Commit a34d730b2 ('Move resolution path elem to ImageBytesR to a query
argument') fixed all the Haskell paths, but forgot to update the JS
viewer too. It was not broken as in not being able to get images, but
it was broken as in always requesting full-size images.

The fix is actually making the imageUrlScaled function much cleaner,
as we don't need to munge the URL itself anymore, but instead can
nicely edit the search string via the URL/URLSearchParams API, which
is well enough supported. So bonus point :)