Minor improvement to SVG error messages
Split into lines and remove the line length restrictions - I still
don't understand the units here, and it was leading to weird effect.

Long term, these should be replaced by HTML elements, but not sure how
yet, or if the server errors should be retrieved at all? They leak
paths/etc, but that's not a big issue anyway (`/settings`), so…