Fix switch to non-loaded images
This can be called the story of three-bugs-with-one-line-fix.

- Switching to prev/next folder was very slow (relatively speaking),
  which I put on the non-cached aspect.

- Switch to prev/next folder resulted in grainy images, which "fixed"
  themselves with switching back and forth, so I thought it's a bug in
  the first-time drawing (on my side).

- Going to "first image" for an invalid image resulted in no output,
  but navigating to same image via "prev" resulted in expected error
  message.

Turns out, all of these were the effect of switchToImage not passing
the resolution query parameter. That means large (slow) images,
large (and then downscale) images, and large/direct loaded images, for
which scaling is not run so the error message is not generated. Fun!