Try to eliminate CI test failures due to session key conflicts
I see this sometimes on Travis:

  Failures:
  …
  uncaught exception: IOException of type ResourceBusy
  config/client_session_key.aes: openBinaryFile: resource busy (file
  is locked)

I believe this is due to the fact that a clean checkout is missing a
session key, and the clientsession library automatically generates a
key if missing. But as tests are run in parallel, multiple tests will
try to generate a key file in parallel, and this leads to the locking
failure, as the writing is done via `writeFile` (from bytestring),
which uses `withBinaryFile` from `System.IO`, which does
open-write-close, not write-to-temp-then-rename.

So the fix is to force a manual "initialise key file" using the same
path as Foundation will later use, then let the tests run normally.