Revert most dependencies to the non-minified version
It appears that yesod does plain concatenation of JS files, which
means that if a file ends with a `//` comment, and no EOL, and the
next file starts with a multi-line comment (`/*`), that will get
ignored, and parsing the combined JS file fails with likely syntax
error.

Filled yesodweb/yesod#1679, in the meantime let's just revert to the
non-minified versions, where easily possible.