Switch login form to hand-written
The login form is created/rendered via Yesod's `generateFormPost` and
`renderBootstrap3`, but the parsing is anyway done in
`authHashDBWithForm` which is using `runInputForm`, so there's not
much gain from using `generateForm`. Let's switch to manually built,
which allows customising the layout easily (`generateForm` also
allows, but in a very verbose way).

This incidentally also reverses the previous commit, since now the
token must be manually added. Oh sigh…