Introduce image open in browser and download actions
This introduces a widget of links for opening the image in the
Corydalis viwer, opening the image the browser (plain view, which
allows zooming as well) and finally downloading the image. Then
switches the image info page and flagged image list to use these
links. The flagged image list also adds a new link for opening the
image info, replacing the previous name/path, since that's duplicate
and more "in your face" (a link that long).