Fix movie info page actions
Currently, the actions list for movies just uses the image actions, which means
trying to view or download a movie will view, respectively download, the
thumbnail for the video, which is not helpful.

Fix this cleanly by adding a couple of abstractions and make sure the actions
widget is consistently populated.

Closes issue #9.