Ensure cache (subdirs) are always created
Otherwise it needs manual intervention for first run or after adding new source
dirs (of either type). This was a not a problem before cache management was
added, since the first exif scanning or rendering would create them, but now
the cache counting is done before any such step gets to run, leading to
breakage.