Add a stop-gap fix for error handling during convert call
When the `convert` binary is not at all available, `proc` will throw an
exception, not simply retun an error code. This kills the entire scanner
thread, rather than simply gathering the errors.

There should be much better handling, but this is a quick fix only
(offline and no access to docs!).