Log rewrite #2: allow silencing request logs
This commit adds a new setting to the settings file, `requests-logging`,
which controls how HTTP requests are logged. Compared to the previous
"detailed-logging", this allows to suppress logs entirely, which is
very useful in tests, as the test output is dominated by various
logs (request, app, SQL, etc.) instead of showing the actual test
results.