Log rewrite #3: log database migrations properly
The default of `runMigration` is to log the migration to stderr, using
plain `hPutStrLn`. This is both very noisy in tests, and also due to
the fact that `hPutStrLn` is not atomic, it leads to completely
mangled output.

Instead, use `runMigrationQuiet`, capture the logs, and log them at
info log level, so that they can be properly filtered or logged as
desired. In the tests config, disable this, finally leading to clean
and quiet hspec-only output!