Implement inline video player for the own viewer
While the fancybox viewer has support for videos since ages, the own
(homegrown) one did not, and simply opened the raw video file in
another tab. This is very user incovenient, and I was planning for a
long time to fix, just was not sure how.

The commit adds supports for vidoes using a simple HTML5 video tag,
running in dynamically instantiated elements, based on the data served
from the backend. It's not the most elegant way, but it works. I tried
to use a static element, but it resulted in sizing problems, and
releasing the element and creating a new one on-demand seems better.

No automated tests, but tested manually. The UI interaction is broken,
will need fixing - the video controls overlap (under) the viewer
controls.