Replace standard movie controls with custom ones
This took a while, but I'm pretty happy with the outcome. The controls
are integrated in the normal navbar, modulo the seek bar, which appears
dynamically when we have a video.

There are a couple of issues, mainly with scrolling before the movie is
ready. Not sure yet whether to hide the bar until the movie is playing,
or to show it but disable the input until the movie is ready.