Switch to format TS/JS code with prettier
This is a large change, as I was _very_ inconsistent with the style,
even mixing single and double quotes on the same line ☹️. So, even real
issues aside, a lot of the lines are change.

The only override for prettier is the single quotes option, as I both
prefer it, and it leads to least delta. For now, only the viewer.ts
file is re-formatted.