Cleanup the homepage itemdivs
Currently, most of the homepage cards have titles that are links, unless
there are no items available, in which case the link is disable, and the
only visual indication is a slight grey image. However, two of the cards
(debug and specialized views) are not buttons, and thus they have a
different padding, leading to visual inconsistency. This is distracting.

Furthermore, the "link-in-title" is not really obvious, and it's
duplicate with the "_(N more)_" link at the bottom of the card. It would
be better to have a single link, and more obvious, so this change:

- drop the link in title, making all card titles simple paragraphs
- moves the "N more" button to the end using a `float-end` class, in the
  hope that it will be more visible, and not lost in the card content
- well, the hope is also that the end alignment is not visually
  distracting in itself…