Add a new homepage card with interesting options
Well, "browse all" and "random" are not very interesting, but the idea
is to add a few more "non-list-type" options to the homepage. The idea
came from the fact that there's no easy way to just look at all images,
or just pick a random image, or any other "spontaneous" action.