Fix browser navigation (backward/forward) in viewer
This was supposed to work, but I have no idea why I was passing `null`
as object to `pushState`, since that defeats the whole purpose. Maybe
I only wanted to record the URL? In any case, using proper state for
push also allows reacting to the `popstate` event, which saves page
reload (and thus is much faster) (I think).