Implement a keyboard shortcuts dialog for the viewer
This is more complex than needed, and I'm not very happy with the outcome.
It needed significant changes to how the modifier keys are handled, and
how the keyboard shortcuts work, as it should really be a "modal" dialog,
but this doesn't work like that in HTML, so it needs faking - basically if
the dialog is shown, any key press should first close. Or not really any
any, modifier keys should be ignored.

However, it seems to work, and it's just a first version. It can be
cleaned up in respect to design, for sure.