Add nicer formatting of filters for e.g. March 15th
Previously this would be "taken in March and on the 15th", which is not
nice. I knew what I wanted, but the scary part is that as soon as I
started typing, Copilot completed the function body (well, it was inline
at first), and when I wanted to add the reverse case (day first), I didn't
write any characters, just pressed enter after the first case, and it
autompleted everything. This is _scary_.