Introduce a new handler for view-from-search
There's no way today to do a search and directly jump into the image
viewer. It is possible to jump into image _browsing_, or image _list_, but
not directly into the image viewer, which means it's always a two step
process. I didn't realise how annoying is this until the "view a random
image", which is oh-so-conveninent, so time to add searcht-to-view.

Full functionality is not yet there, but this is the first half or so.