Introduce the "direct view" mode
This is allows staying in image view mode after new searches (good),
but is activated always on entering the image view mode (rather bad). I
tried to make it only on selecting view mode explicitly (via the new
SearchViewR route), but it didn't work out, as either Yesod doesn't send
cookies on redirect, or Safari doesn't process them. In any case, to
make it work, the only way I saw was to enable it on ViewR, so there it
is. Will likely get tweaked in the future.