Viewer: reorg menu and move folder actions into it
I think the folder backwards/forwards were not used often enough to
warrant a position in the image nav bar, so let's move them into the new
menu, and clean up the ordering of the menu items too.

Moving the folder navigation into the menu also requires dropping the
movie forward/backward buttons, as they have no pairing, but that's ok -
there was no keyboard shortcut for them, and if you use the mouse or touch
interface anyway, you can use the scroll bar directly, which is cleaner -
the fixed ten second step was not appropriate for all movie lenghts.