Disable zooming by double tapping on the navbar
Since the viewer interface is managed (sizing/zooming), enabling tap to
zoom is actually confusing. Just learned about disabling this via the
`touch-action` property, so let's do it.

I think for viewer it would be better to disable page-wide, but then
zooming on the search box (the only way I can trigger it) is not that
bad, since it's hard to accidentally trigger.