Stop manually linking to scripts in curate pages
The curate pages (lens, camera, main curate page) were manually linking
to the scripts, which requires deploying the individual JS files in
production. To simplify production deployments, add all these small
scripts to the `PagePlot` route type (they are small and the overhead in
bytes is smaller, I think, than the extra resource overhead).

This allows deploying the `static` directory just with the `combined`
subdir, and the webfonts (which can't be combined with other resources).