Fix document loaded conflicts for the curate pages
In the merging the individual JS scripts in 56fcb3c, I introduced a bug
that was not clearly obvious (and there are no tests for this). Basically,
all scripts were using document loaded events, which means all scripts
were firing for all pages, leading to conflicts since the resources were
not compatible (but not completely independent either, due to reuse of div
IDs across pages). This resulted in empty Plotly graphs.

To fix this, make the scripts just define the functions, and then
install the event in the actual hamlet template.