viewer: Simplify drawing transformation/drawing offsets
For some reason lost to history, the code was doing:

- compute offsets
- apply transformation matrix shifting each axis by A and B
- subtract A from offset X and B from offset Y
- call draw with the new offsets X and Y

Since the transformation matrix offsets are additive to the draw call
parameters, the smaller offsets are nullifying the transform offsets.

So the entire thing can be simplified by keeping the transformation
matrix only for (potential) rotation, and doing the draw offsets as
originally computed - for centering the image on the borders which are
smaller than the canvas.