Fix a long-standing bug in folder listing with some atoms
Some atoms, like folder-class, do not result in images - they can only
find folders. The folder listing always generates thumbnails (for the
found folders), and the thumbnails were statically linking to the image
providing the thumbnail, together with the search params, which would
result in "no such image" 404. This might have been the case of the
bot-generated 404s, with valid referrer values, that confused me for
many years.

This change fixes it to dymamically link to either an image or to its
parent folder, depending on atom type. The result code is not nice, and
can be cleaned further, but this is the direct change for doing this
conditional aspect.