viewer: Implement proper panning!
Finally, after 20 versions, this works. AI didn't help, as it wanted me
to go down a rabbit hole of complex panning computations and clamping
limnits, so I actually went away from the keyboard and thought how to
simplify it. And yes, the final version is ~5 lines of code, with a
_very_ simple model. But I added way more comment lines, to make sure
I don't forget _again_ how this is intended to work.