viewer: Initial support for pinch-zoom (touch)
This was originally machine-generated, but then I changed the code flow.
There are two missing parts: one, the pinch-zoom is screen-centered, when it
should be touch-point-centered. Second, there's no pan support, for
which the small complexity is converting absolute pixel movements into
relative ones - this will need integration with the drawImage, which for
a clean API, will require also storing the canvas size and overflow
values into corydalis state, which will allow much better keyboard
panning support for corner cases.